+++ lib/bolshoi.dart
@@ -3,6 +3,7 @@ library bolshoi;
 export 'src/animation_helpers.dart';
 export 'src/property_animation.dart';
 export 'src/animation_sequence.dart';
+export 'src/animation_group.dart';
 export 'src/tweens.dart';
 export 'src/i_animation.dart';