+++ lib/src/animation_helpers.dart
@@ -1,8 +1,10 @@
 import 'dart:async';
 import 'dart:collection';
 
+import 'package:bolshoi/bolshoi.dart';
 import 'package:bolshoi/src/property_animation.dart';
 import 'package:flutter/animation.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
 const kDuration = const Duration(milliseconds: 500);
@@ -10,11 +12,43 @@ const kDuration_ms = 500;
 const kDuration2 = const Duration(milliseconds: 360);
 const kDuration2_ms = 360;
 
-/// animation group
+/// opacity animation helper
+class Fade extends Animate<double> implements IAnimation {
+  Fade(int milliseconds,
+      {@required OnTick<double> onTick,
+      @required TickerProvider ticker,
+      Curve curve = Curves.easeIn,
+      bool out: false})
+      : super(
+          milliseconds,
+          curve: curve,
+          onTick: onTick,
+          begin: out ? 1.0 : 0.0,
+          end: out ? 0.0 : 1.0,
+          ticker: ticker,
+        );
+}
+
+///
+class Zoom extends Animate<double> implements IAnimation {
+  Zoom(int milliseconds,
+      {OnTick animator, TickerProvider vsync, bool out: false})
+      : super(
+          milliseconds,
+          curve: Curves.easeIn,
+          onTick: animator,
+          begin: out ? 1.0 : 0.0,
+          end: out ? 0.0 : 1.0,
+          ticker: vsync,
+        );
+}
+
+/// Deprecated animation queue => vf. cf [AnimationSequence]
 ///
 /// future : base on bi-directionnal queue ?
 /// [BidirectionalIterator] [cf. doc](https://api.dartlang.org/stable/1.22.1/dart-core/BidirectionalIterator-class.html)
 ///
+@deprecated
 class AnimationQueue extends ListQueue<PropertyAnimationBase> {
   AnimationStatus currentStatus = AnimationStatus.dismissed;