+++ test/animation_queue_test.dart
@@ -1,16 +1,18 @@
 import 'package:bolshoi/bolshoi.dart';
+import 'package:flutter/animation.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('Animation queue basics', () {
-    AnimationQueue queue;
+  group('Animation group basics', () {
+    AnimationGroup anims;
 
     setUp(() {
-      queue = new AnimationQueue();
+      anims = new AnimationGroup([]);
     });
     test('new instance has a listenable status', () {
-      queue.statu$.listen((s) => s);
-      expect(queue.statu$ != null, equals(true));
+      anims.statu$.listen((s) => s);
+      expect(anims.statu$ != null, equals(true));
+      expect(anims.statu$ , equals(AnimationStatus.dismissed));
     });
   });
 }